Use skip_empty_rows in call to read_fwf
In readr 1.4.0 read_fwf had a bug that caused `skip_empty_rows` not to
function properly when the rows had just whitespace. In readr 2.0.0 this
works as intended, which means if you would prefer the existing behavior
we would need to set this option to false in order for the tests to
pass.