Fix issue with escape characters in strings.
The problem was due to format.info() being used to determine the width
of strings. Unfortunately, format.info() includes escape characters in
the returned width, leading to width values that do not match the number
of characters written. This causes problems when writing strings with
escape characters in fixed width format.

The issue is solved by replacing format.info() with nchar(), wherever
strings are concerned.